fixing up PR comments
